Add Instructions & BLE NimBLE Fix
Fixed a bunch of small issues, like formatting, typos, and inefficient variable space uses in `MyDefaults_struct`.

Added instructions on what libraries are necessary to use this software.

Fixed issue where bluetooth ESP32 device would not reconnect or re-advertise itself to the phone to allow key presses to be sent when the turntable is powered off after device is paired. Only first connection after a fresh pairing with the phone would allow the commands to be sent correctly. This caused issues that was previously only fixed by removing the turntable as a bluetooth device from the phone and pairing again.

Credit to https://github.com/T-vK/ESP32-BLE-Keyboard/issues/247 and https://forum.arduino.cc/t/esp32-ble-does-not-disconnect/957235/4 for helping me solve this.

Added instructions to fix the bluetooth issue by installing NimBLE-Arduino 1.4.1+ and installing ESP32-BLE-Keyboard v0.3.2-beta (which is over a year old at this point) to enable the use of NimBLE, which allows the automatic pairing and reconnecting of the bluetooth device.